# backend/src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payoova Wallet API",
    "description": "Professional cryptocurrency wallet API with multi-network support",
    "version": "2.0.0",
    "contact": {
      "name": "Payoova Support",
      "email": "support@payoova.com",
      "url": "https://payoova.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://qjh9iec79z56.manus.space/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/kyc/status": {
      "get": {
        "summary": "Get user KYC status",
        "tags": ["KYC"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "KYC status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KYCStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/kyc/verify": {
      "post": {
        "summary": "Submit KYC verification",
        "tags": ["KYC"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KYCVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "KYC verification submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KYCVerificationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "KYCStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "status": {
            "type": "string",
            "enum": ["pending", "approved", "rejected", "not_started"]
          },
          "verification_level": {
            "type": "integer"
          },
          "documents_required": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "KYCVerificationRequest": {
        "type": "object",
        "required": ["first_name", "last_name", "date_of_birth", "country"],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date"
          },
          "nationality": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "address_line1": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postal_code": {
            "type": "string"
          }
        }
      },
      "KYCVerificationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "default": true
          },
          "status": {
            "type": "string"
          },
          "verification_id": {
            "type": "integer"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}